moved signout button to seperate component
